requirements.txt
# ============================================================================
# Agent Attack Framework — Full Dependency Specification
# ============================================================================
#
# Target: Python 3.11  |  CUDA 12.8  |  RTX 5090 (32 GB)
#
# SETUP INSTRUCTIONS (order matters due to JAX / PyTorch coexistence):
#
#   1. Create conda environment:
#        conda create -n libero python=3.11 -y
#        conda activate libero
#
#   2. Install C compiler (needed by Triton / vLLM CUDA graph compilation):
#        conda install -c conda-forge gcc_linux-64 gxx_linux-64 -y
#        # Then create symlinks in $CONDA_PREFIX/bin:
#        ln -sf $CONDA_PREFIX/bin/x86_64-conda-linux-gnu-gcc $CONDA_PREFIX/bin/gcc
#        ln -sf $CONDA_PREFIX/bin/x86_64-conda-linux-gnu-g++ $CONDA_PREFIX/bin/g++
#        ln -sf $CONDA_PREFIX/bin/gcc $CONDA_PREFIX/bin/cc
#
#   3. Install OpenGL libs (for MuJoCo headless rendering):
#        conda install -c conda-forge libopengl mesalib -y
#
#   4. Install PyTorch (CUDA 12.8):
#        pip install torch==2.9.0 torchvision==0.24.0 torchaudio==2.9.0 \
#            --index-url https://download.pytorch.org/whl/cu128
#
#   5. Install pinned JAX + numpy (must come before openpi):
#        pip install "jax[cuda12]==0.5.3" "numpy>=1.26,<2.0"
#
#   6. Install this requirements file:
#        pip install -r requirements.txt
#
#   7. Install local editable packages (adjust paths to your machine):
#        pip install -e /path/to/LIBERO    --no-deps
#        pip install -e /path/to/openpi    --no-deps
#        pip install    openpi-client      --no-deps
#
#   8. (Optional) Set env vars in your .env or shell:
#        CC=$CONDA_PREFIX/bin/gcc
#        XLA_PYTHON_CLIENT_MEM_FRACTION=0.25
#
# ============================================================================

# ---------------------------------------------------------------------------
# 1. Core RL / training framework
# ---------------------------------------------------------------------------
openpipe-art[backend,langgraph]==0.5.9
vllm==0.13.0
trl==0.24.0
unsloth==2026.2.1
transformers==4.57.6

# ---------------------------------------------------------------------------
# 2. Agent orchestration (LangGraph / LangChain)
# ---------------------------------------------------------------------------
langgraph==1.0.8
langchain-core==1.2.12

# ---------------------------------------------------------------------------
# 3. LLM clients & observability
# ---------------------------------------------------------------------------
openai==2.21.0
litellm==1.81.11
weave==0.52.28

# ---------------------------------------------------------------------------
# 4. JAX ecosystem (for Pi0.5 VLA inference via OpenPI)
#    NOTE: jax[cuda12] and numpy are pinned in step 5 above.
# ---------------------------------------------------------------------------
jaxlib==0.5.3
flax==0.10.2
equinox==0.13.4
orbax-checkpoint            # version resolved by flax
chex==0.1.86
jaxtyping==0.2.36
augmax==0.4.1

# ---------------------------------------------------------------------------
# 5. Robotics / simulation (MuJoCo, robosuite, LIBERO deps)
# ---------------------------------------------------------------------------
mujoco==3.4.0
robosuite==1.4.0
bddl==3.6.0
easydict==1.13
h5py==3.15.1
PyOpenGL==3.1.10
PyOpenGL-accelerate==3.1.10
termcolor==3.3.0

# ---------------------------------------------------------------------------
# 6. ML / scientific computing
# ---------------------------------------------------------------------------
numpy>=1.26,<2.0
scipy==1.17.0
scikit-learn==1.8.0
scikit-image==0.26.0
opencv-python==4.13.0.92
Pillow==12.1.0
imageio==2.37.2

# ---------------------------------------------------------------------------
# 7. Tokenizers / model utilities
# ---------------------------------------------------------------------------
sentencepiece==0.2.1
einops==0.8.2
dm-tree==0.1.9
beartype==0.19.0
tyro==1.0.6
ml-collections==1.0.0
numpydantic==1.7.0
treescope==0.1.10
tqdm-loggable==0.3
flatbuffers==25.12.19

# ---------------------------------------------------------------------------
# 8. General utilities
# ---------------------------------------------------------------------------
pydantic==2.12.5
python-dotenv==1.2.1
httpx==0.28.1
tqdm==4.67.3
packaging==25.0
nest-asyncio==1.6.0
hydra-core==1.3.2
omegaconf==2.3.0
